// src/app/models/personalidad.model.ts
export interface Personalidad {
    id: string

    detalles: string
    imagen: string
    
    mbti: string;
    colorForma: string;
    oficio: string;
    descripcion: string;
    comoSon: string;
    cualidad: string;
    dondeEncaja: string;
    habilidad: string,
    valores: string;
    pareja: string;
    traumas: string;
    enfermedad: string;
    danan: string;
    CaleVRije: string;
    funcion: string;
    grupo: string;

    updatedAt: number;
    _deleted: boolean;
}